+++ packages/terra-form-textarea/tests/nightwatch/AutoResizableTextarea.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import Textarea from '../../lib/Textarea';
+
+const textarea = () => <Textarea id="auto-resizable" cols="2" isAutoResizable defaultValue="Default Value" />;
+
+export default textarea;